requirements.txt
streamlit
openai
langchain
chromadb
sentence-transformers
python-dotenv
tqdm
requests
pypdf